[Website] Prevent the initial flash of "You have no Playgrounds" message
Sometimes when Playground is loaded, the user sees a brief flash of the
"You don't have any Playgrounds yet" message. This PR fixes that. The
user will see the spinner until their Playground loads.

 ## Implementation

Before this PR, the message is displayed when:

 * The OPFS sites are already loaded
 * The temporary site was not created yet in the redux store

With this PR, the message will never be displayed again. It's removed
entirely. There should never be a situation where a fully loaded
Playground has no sites. The temporary site is always there.

Furthermore, this PR keeps track of the temporary site initialization state and
flips a boolean flag called `setFirstTemporarySiteCreated` once the
first temp site is ready.

 ## Testing instructions

This one is tricky to test as the bug relies on a specific race
condition. You might add `await new Promise(resolve=>setTimeout(resolve,
10000));` to setTemporarySiteSpec to delay that part and confirm you
only see the spinner for 10 seconds.