Make cookie store web-only
It is needed on the web where custom Response objects
may not be assigned Set-Cookie headers. Otherwise, cookies
from PHP requests would not be respected.
But a cookie store isn't needed for Playground CLI,
and applying the same cookies to all requests to a CLI-based
server interferes with clients that want to use other
WP auth schemes.